+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata, Viewport } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
-import { PrimeReactProvider } from 'primereact/api'
+import { AppProvider } from '@/lib/providers'
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -28,7 +28,7 @@ export default function RootLayout({
 	return (
 		<html lang="en" className="dark">
 			<body className={inter.className}>
-				<PrimeReactProvider>{children}</PrimeReactProvider>
+				<AppProvider>{children}</AppProvider>
 			</body>
 		</html>
 	)