+++ src/lib/providers/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Provider exports for KAI application
+ *
+ * @description
+ * Centralized exports for all application providers.
+ * Import from this file to get access to all providers.
+ */
+
+export { PrimeReactProvider } from './PrimeReactProvider'
+export { KaiThemeProvider } from './KaiThemeProvider'
+export { AppProvider } from './AppProvider'